[ci] Cache playwright in run_devtools_e2e_tests
I happened to notice that I forgot to cache playwright in run_devtools_e2e_tests, so it would try to install it every time which can randomly take a while to complete (I'm not sure why it's not deterministic, but the dependencies appear to be installed inconsistently across multiple workflows).

This PR adds the same cache we use for other steps that use playwright, which should shave off some time from this workflow when the cache is warm.