[ci] Cache playwright in run_devtools_e2e_tests
I happened to notice that I forgot to cache playwright in run_devtools_e2e_tests, so it would try to install it every time which can randomly take a while to complete (I'm not sure why it's not deterministic, but the dependencies appear to be installed inconsistently across multiple workflows).

This PR adds the same cache we use for other steps that use playwright, which should shave off some time from this workflow when the cache is warm.

Additionally I omitted the standalone install-deps command as it appears to be redundant and adds a lot of extra time to CI, due to the fact that it installs many unrelated dependencies.